requirements.txt
absl-py==2.0.0
aenum==3.1.15
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
dm-tree==0.1.8
Flask==3.0.0
frozenlist==1.4.1
future==0.18.3
google-auth==2.25.2
google-auth-oauthlib==1.2.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
keras==3.0.2
line-bot-sdk==3.6.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
namex==0.0.7
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
Pillow==10.1.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rsa==4.9
six==1.16.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
typing_extensions==4.9.0
urllib3==2.1.0
v==1
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
